// _app-mobile/src/modules/sales/store/sales.store.types.ts
export type TFocusedProps = 'period' | 'product' | 'offer';

export type PaymentMethodProps = 'creditCard' | 'bankSlip' | 'pix' | 'free';

export type TFilterProps =
  | 'approved'
  | 'pending'
  | 'expired'
  | 'refused'
  | 'refunded'
  | 'chargeback'
  | 'utmCampaign'
  | 'utmContent'
  | 'utmMedium'
  | 'utmSource'
  | 'utmSrc'
  | 'utmTerm'
  | 'creditCard'
  | 'bankSlip'
  | 'pix'
  | 'free';